Fix four memory-safety bugs and restore the creature-name combo
Reported by a downstream user running LibreCreatures under AddressSanitizer;
each checked against the native binary before fixing.

- Macro holder double free: a started, non-capturing script that ends marks
  itself destroy-when-finished and deletes itself, but its MacroHolder kept
  the pointer and released it again on destruction. A Macro now carries its
  owning holder and clears the holder's pointer when destroyed; the holder
  adopts a fresh Macro if a kit reuses it. The pipe server, whose holder
  dies as soon as the command returns, now hands a still-running started
  Macro to the scheduler -- previously the holder's destructor unscheduled
  the script it had just started, so scheduled pipe scripts never got past
  their first wait. (Native keeps the pointer too; its kits' holders outlive
  the script.)
- Bacterium input chemical: ids are -8..-1 as signed bytes (chemicals
  248-255) and were widened signed into add_chemical_moles, writing up to
  eight entries before chemical_states. UpdateBacteriumAndEnvironment
  @ 0x00409190 zero-extends.
- Interaction pose table: indexed up to 23 but declared with 22 rows. The
  native table @ 0x0045abc8 has 24; the two missing rows, ('2','1') and
  ('3','1'), are restored from the binary.
- A creature killed inside its own script purged the Macro still executing
  that command and the interpreter ran on through freed memory. The purge
  now only unschedules and marks an executing Macro; the interpreter deletes
  it when the command returns.
- PointerTool::SetTextWithToolbarUpdate @ 0x00429860 moves the text to the
  top of the toolbar combo's ten-entry history; the port rebuilt the
  creature menu instead, so Speak (Ctrl+S) had nothing to say.

+++ src/c1/creatures/creature.cpp
@@ -3573,9 +3573,14 @@ void Creature::update_bacterium_and_environment(
                 common::debug_log(*log_host, 0x400, "Bacterium killed!\n");
             }
         } else {
+            // UpdateBacteriumAndEnvironment @ 0x00409190 passes the id
+            // zero-extended.  Bacterial input ids are -8..-1 as signed bytes
+            // (chemicals 248-255); widening them signed indexed up to eight
+            // entries before chemical_states.
             biochemistry.add_chemical_moles(
-                static_cast<int>(bacterium.input_chemical_id()), 0x50,
-                log_host);
+                static_cast<int>(
+                    static_cast<std::uint8_t>(bacterium.input_chemical_id())),
+                0x50, log_host);
             for (const std::int8_t output_chemical_id :
                  bacterium.output_chemical_ids()) {
                 if (output_chemical_id != 0) {
